+++ packages/vite-plugin-astro/src/index.ts
@@ -1,6 +1,10 @@
 export { default } from './plugin/index.js';
 export { getAstroMetadata, createDefaultAstroMetadata } from './plugin/metadata.js';
 export type { AstroPluginMetadata } from './plugin/types.js';
-export type { CompilerError, CSSError, AggregateError } from './types.js';
-export type { CompileProps, CompileResult } from './compile/index.js';
-export { compile } from './compile/index.js';
+export type {
+	CompilerError,
+	CSSError,
+	AggregateError,
+	AstroPluginOptions,
+	ExposedTransformOptions,
+} from './types.js';
